🔄 synced local 'skyvern/' with remote 'skyvern/'
<!-- ELLIPSIS_HIDDEN -->

> [!IMPORTANT]
> Migrate `Organization` model from `models.py` to `schemas/organizations.py` and update imports accordingly.
>
>   - **Imports**:
>     - Update import paths for `Organization` from `skyvern/forge/sdk/models` to `skyvern/forge/sdk/schemas/organizations` in `agent_functions.py`, `app.py`, and `cloud_permission_checker.py`.
>     - Similar updates in `db_client.py`, `checkout.py`, and `observer.py`.
>   - **Models**:
>     - Remove `Organization` and `OrganizationAuthToken` classes from `models.py`.
>     - Add `Organization` and `OrganizationAuthToken` classes to `schemas/organizations.py`.
>   - **Misc**:
>     - Adjust related imports in `agent.py`, `agent_functions.py`, and `org_auth_service.py`.
>
> <sup>This description was created by </sup>[<img alt="Ellipsis" src="https://img.shields.io/badge/Ellipsis-blue?color=175173">](https://www.ellipsis.dev?ref=Skyvern-AI%2Fskyvern-cloud&utm_source=github&utm_medium=referral)<sup> for ea501c5324f88a81c7abf3a1ea5f44ae8d821fa9. It will automatically update as commits are pushed.</sup>

<!-- ELLIPSIS_HIDDEN -->
