+++ skyvern/forge/sdk/services/org_auth_token_service.py
@@ -4,7 +4,7 @@ import structlog
 
 from skyvern.forge.app import DATABASE
 from skyvern.forge.sdk.core import security
-from skyvern.forge.sdk.models import OrganizationAuthToken, OrganizationAuthTokenType
+from skyvern.forge.sdk.schemas.organizations import OrganizationAuthToken, OrganizationAuthTokenType
 
 LOG = structlog.get_logger()
 API_KEY_LIFETIME = timedelta(weeks=5200)
